unsigned long to unsigned long long

+++ src/modular_operations/modular_addition.c
@@ -21,13 +21,13 @@ inline void modular_addition(const Uint256 *a, const Uint256 *b, Uint256 *result
   // SO: subtract p when outside secp256k1 space OR overflowFlag is true
 
   // TODO: create a function to do this. modulus and this (and probably others will) use this
-  unsigned long isOutsideSecp256k1Space = 0;
+  unsigned long long isOutsideSecp256k1Space = 0;
   isOutsideSecp256k1Space |= (tmp.limbs[0] > SECP256K1_P_0);
   isOutsideSecp256k1Space |= ((tmp.limbs[0] == SECP256K1_P_0) & (tmp.limbs[1] > SECP256K1_P_1));
   isOutsideSecp256k1Space |= ((tmp.limbs[0] == SECP256K1_P_0) & (tmp.limbs[1] == SECP256K1_P_1) & (tmp.limbs[2] > SECP256K1_P_2));
   isOutsideSecp256k1Space |= ((tmp.limbs[0] == SECP256K1_P_0) & (tmp.limbs[1] == SECP256K1_P_1) & (tmp.limbs[2] == SECP256K1_P_2) & (tmp.limbs[3] >= SECP256K1_P_3));
 
-  const unsigned long toSubtractMask = -(isOutsideSecp256k1Space | ((unsigned long) overflowFlag));
+  const unsigned long long toSubtractMask = -(isOutsideSecp256k1Space | ((unsigned long long) overflowFlag));
 
   const Uint256 toSubtract = (Uint256){.limbs = {
                                            SECP256K1_P_0 & toSubtractMask,
